introduce handleOrganizationNameUpdateSaga + test
